requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
fastapi==0.109.0
fonttools==4.47.2
h11==0.14.0
idna==3.6
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.1
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.4
uvicorn==0.27.0
